terrabit_ats_full_branded/ats_app_branded/requirements.txt
streamlit>=1.36.0
openai>=1.30.0
pymupdf>=1.24.7
python-docx>=1.1.2
pandas>=2.2.2
openpyxl>=3.1.2
spacy>=3.7.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
audio-recorder-streamlit>=0.0.8
reportlab>=4.2.2
